Blog list frontend, step6
